pep-8 nos nomes das variaveis, reescrevendo os testes, adicionando atributo ao cog
